Adding tests and code for short-circuit execution
Sometimes the actions taken by the sub-type don't allow for pagination or aren't
really for examining data (has a user starred something or not - would like that
to return a boolean). So, now there's a way to short-circuit the process.

Old Process (still works):
    1. Call the type to prep for the action
    2. Call the action on the type (returns github_request)
    3. Manipulate github_request object if you want to (ex. Pagination)
    4. Execute the request getting a github_result object

Can short-circuit at step 2 - having it return useful information there. If the
sub-type action returns a staticarray, the first item being the http_request,
and the next item being either a boolean (true) to return a the github_result
right away, or a capture with code to run on the result of issuing the request.
(The capture gets passed a github_result as the first argument when invoking.)